Add option to free memory manually to avoid OOM kill
On low-memory systems (e.g. NAS) Go's GC may not run frequently enough,
resulting in high memory usage that gets the process OOM killed. This
option allows specifying an interval to trigger GC and free memory back
to the OS. When combined with the verbose option, memory usage stats
will also be printed to show the change in memory use.