Handle selectively synced local copy
Add `--selective` flag to enable selective sync validation. In this
mode, we assume that the top-level folders present in the local root are
the only ones that are meant to be synced, and filter the remote results
accordingly. This does mean that any files at the first level of the
local root will not be validated, only ones contained in a subfolder.